cto: add default mmap hint
When the pmemcto pool is created and there is no mmap hint specified,
the system is allowed to pick the first available address range to place
the pool mapping.  If there are some differences between 'create' and
'open' paths in the program and there are any additional allocations
made before the pool is open for the next time (i.e. there are some
threads spawned, etc.), the base pool address may conflict with those
allocations and the library would fail to open the pool.
On Windows, due to ASLR, the problem may occur on subsequent program
runs.

To make it less likely, libpmemcto will use the default hint address for
mmap (initially hardcoded to 0x10000000000) to avoid conflicts with
the heap.  It can still be overriden via PMEM_MMAP_HINT.